requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
idna==3.10
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
markupsafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
nats-py==2.9.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
setuptools<81
streamz==0.6.4
sympy==1.13.1
threadpoolctl==3.6.0
toolz==1.0.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
triton==3.2.0
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
zict==3.0.0
nfstream==6.5.3
